.old/shards/requirements.txt
sqlalchemy
pyyaml
watchdog  # For future file monitoring
python-dotenv  # For configuration